Move server related classes from src/test to src/main
