+++ src/main/java/org/springframework/reactive/web/http/reactor/ReactorHttpServer.java
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.reactive.web.http;
+package org.springframework.reactive.web.http.reactor;
 
 import org.springframework.beans.factory.InitializingBean;
+import org.springframework.reactive.web.http.HttpServer;
+import org.springframework.reactive.web.http.HttpServerSupport;
 import org.springframework.reactive.web.http.reactor.RequestHandlerAdapter;
 import org.springframework.util.Assert;
 import reactor.bus.selector.Selectors;
