+++ src/main/java/org/springframework/reactive/web/http/rxnetty/RxNettyHttpServer.java
@@ -13,11 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.reactive.web.http;
+package org.springframework.reactive.web.http.rxnetty;
 
 import io.netty.buffer.ByteBuf;
 
 import org.springframework.beans.factory.InitializingBean;
+import org.springframework.reactive.web.http.HttpServer;
+import org.springframework.reactive.web.http.HttpServerSupport;
 import org.springframework.reactive.web.http.rxnetty.RequestHandlerAdapter;
 import org.springframework.util.Assert;
