+++ src/main/java/org/springframework/reactive/web/http/servlet/JettyHttpServer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.reactive.web.http;
+package org.springframework.reactive.web.http.servlet;
 
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.ServerConnector;
@@ -21,7 +21,8 @@ import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
 
 import org.springframework.beans.factory.InitializingBean;
-import org.springframework.reactive.web.http.servlet.HttpHandlerServlet;
+import org.springframework.reactive.web.http.HttpServer;
+import org.springframework.reactive.web.http.HttpServerSupport;
 import org.springframework.util.Assert;
 import org.springframework.util.SocketUtils;
