+++ src/test/java/org/springframework/reactive/web/http/AbstractHttpHandlerIntegrationTests.java
@@ -21,6 +21,10 @@ import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
+import org.springframework.reactive.web.http.reactor.ReactorHttpServer;
+import org.springframework.reactive.web.http.rxnetty.RxNettyHttpServer;
+import org.springframework.reactive.web.http.servlet.JettyHttpServer;
+import org.springframework.reactive.web.http.servlet.TomcatHttpServer;
 import org.springframework.util.SocketUtils;
